# fixed the bug that searchKeys would return value != -1 when the keys …