--- a/parser_test.go
+++ b/parser_test.go
@@ -261,6 +261,13 @@ var getTests = []GetTest{
 		path:    []string{"a", "c"},
 		isFound: false,
 	},
+	GetTest{
+		desc:    "handling nested json",
+		json:    `{"a":{"b":{"c":1},"d":4}}`,
+		path:    []string{"a", "d"},
+		isFound: true,
+		data:    `4`,
+	},
 
 	// Error/invalid tests
 	GetTest{