**Description**: What this PR does
fixed the bug that searchKeys would return value != -1 when the keys located in different json entry 

For example, trying to search keys ["a", "c"] on following JSON should be failed: 
```json
{
    "a": {
        "a": 1
    },
    "b": {
        "a": 3,
        "c": [1, 2]
    }
}
```


**Benchmark before change**:
BenchmarkJsonParserLarge-4                    	  100000	     72548 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	  500000	     11408 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6554 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7190 ns/op	     544 B/op	      11 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	  500000	     11854 ns/op	     608 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       993 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       634 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	      1047 ns/op	     176 B/op	       7 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       975 ns/op	     240 B/op	       8 allocs/op
PASS
ok  	jsonparser/benchmark	74.007s

**Benchmark after change**:
BenchmarkJsonParserLarge-4                    	  100000	     72744 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	  500000	     13456 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6479 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7017 ns/op	     544 B/op	      11 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	  500000	     11701 ns/op	     608 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	      1002 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       626 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	      1064 ns/op	     176 B/op	       7 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	      1037 ns/op	     240 B/op	       8 allocs/op
PASS
ok  	jsonparser/benchmark	75.685s

For running benchmarks use:
```
go test -test.benchmem -bench JsonParser ./benchmark/ -benchtime 5s -v
# OR
make bench (runs inside docker)
```
…located in different json entry

For example, this json object, when looking for keys [‘a’, ‘c’],
searchKeys should return a negative value, use to return possible which
could cause wrong data for Get API
{
    "a": {
        "a": 1
    },
    "b": {
        "a": 3,
        "c": [1, 2]
    }
}