@buger 2016-12-16
Thank you! Can you explain bug by itself. Why your fix is working?