@onepill 2016-12-16
@buger Hi, you are welcome, thank you for this awesome json library.
And for the bug, here is how it occurs in my understanding
let's take this for example
```json
{
    "a": {
        "a": 1
    },
    "b": {
        "a": 3,
        "c": [1, 2]
    }
}
```
When calling searchKeys on this data with `keys` ["a", "c"], the function will walk thru every character of json string, and whenever found a matching key, keyLevel will be increased.
In our case, when it found key "a", keyLevel would be 1 and level is also.
And it would continue on the remaining data, when hitting the "{" of entry "a", level will be 2, since there is no entry with key "c" in it, it will hit "}", the level will be reduced to 1. (meanwhile keyLevel is still 1)
And when it hits entry "b", level would be 2, and in this entry, it will find key "c", then keyLevel will be added to 2(which is wrong), which is same as the length of parameter `keys`. the function would return a value greater than 1.
So the bug is, searchKeys would return value greater than 1(means found keys) on this JSON data with keys["a", "c"]. which should return a negative value with.