@onepill 2016-12-16
Sorry, found a situation which will be failed due to this change, will add more tests, and reopen this pull request after i fixed it.